backend/requirements.txt
uagents
uagents-core
openai>=1.0.0
python-dotenv
